SVN revision 13056

+++ src/common/misc/server-config.h
@@ -232,7 +232,7 @@ typedef struct server_configuration_s
     int db_cache_size_bytes;      /* cache size to use in berkeley db */
                                   /*   if zero, use defaults */
     char *db_cache_type;
-    int db_max_size;              /* size of database map
+    long db_max_size;             /* size of database map
                                    */
     int trove_alt_aio_mode;       /* enables experimental alternative AIO
                                    * implementation for some types of
